requirements.txt
Pillow
pokinator
guizero
